Initialised favicon edits, configured eslint
