+++ package.json
@@ -29,8 +29,13 @@
     "@types/react": "^17.0.38",
     "@types/react-dom": "^17.0.11",
     "@types/three": "^0.136.1",
-    "eslint": "8.6.0",
+    "eslint": "^8.42.0",
+    "eslint-config-airbnb": "^19.0.4",
     "eslint-config-next": "12.0.7",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "file-loader": "^6.2.0",
     "typescript": "^4.5.4"
   }
