--- pages/index.js
@@ -1,28 +1,0 @@
-import Head from 'next/head'
-import styles from '../styles/Home.module.css'
-import Landing from '../components/Landing'
-import RippleScene from '../components/threejs/LandingScene'
-
-export default function Home() {
-  return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      
-      <RippleScene which="particle" style={{ height: "100vh", width: "100vw", position: "absolute", zIndex: -10 }} />
-      <Landing /> 
-
-      <main role="main">
-        {/* <About />
-        <HowItWorks /> */}
-        {/* <Pricing />
-        <JoinNow /> */}
-      </main>
-
-      {/* <Footer /> */}
-    </div>
-  )
-}
