+++ pages/index.jsx
@@ -1,0 +1,38 @@
+import React from 'react';
+import Head from 'next/head';
+import styles from '../styles/Home.module.css';
+import Landing from '../components/Landing';
+import RippleScene from '../components/threejs/LandingScene';
+
+export default function Home() {
+  return (
+    <div className={styles.container}>
+      <Head>
+        <title>Darren Wong</title>
+        <meta name="description" content="Darren Wong's portfolio" />
+        <link rel="icon" href="/assets/favicon/favicon.ico" />
+        <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
+        <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
+        <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
+        <link rel="manifest" href="/site.webmanifest" />
+      </Head>
+
+      <RippleScene
+        which="particle"
+        style={{
+          height: '100vh', width: '100vw', position: 'absolute', zIndex: -10,
+        }}
+      />
+      <Landing />
+
+      <main role="main">
+        {/* <About />
+        <HowItWorks /> */}
+        {/* <Pricing />
+        <JoinNow /> */}
+      </main>
+
+      {/* <Footer /> */}
+    </div>
+  );
+}
